+++ app/src/main/java/org/supla/android/db/ElectricityMeasurementItem.java
@@ -199,7 +199,7 @@ public class ElectricityMeasurementItem extends IncrementalMeasurementItem {
         Complement = cursor.getInt(cursor.getColumnIndex(
                 SuplaContract.ElectricityMeterLogEntry.COLUMN_NAME_COMPLEMENT)) > 0;
 
-        setProfileId(cursor.getInt(cursor.getColumnIndex(SuplaContract.ElectricityMeterLogEntry.COLUMN_NAME_PROFILEID)));
+        setProfileId(cursor.getLong(cursor.getColumnIndex(SuplaContract.ElectricityMeterLogEntry.COLUMN_NAME_PROFILEID)));
         
     }
